+++ src/main/java/cart/dao/dummyData/MemberInitializer.java
@@ -1,5 +1,6 @@
-package cart.dao;
+package cart.dao.dummyData;
 
+import cart.dao.MemberDao;
 import cart.entity.MemberEntity;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
