+++ src/test/java/cart/dao/JdbcMemberDaoTest.java
@@ -1,7 +1,7 @@
 package cart.dao;
 
+import cart.dao.dummyData.MemberInitializer;
 import cart.entity.MemberEntity;
-import cart.entity.ProductEntity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
